refactor: merge main_http_test.go into main_test.go
